Updated the cupy_type using libs. so they take into account the CudaHost case.

+++ hopfield_dictionary/host_cpp_lib/hopfield_dict.cpp
@@ -7,7 +7,7 @@
 // matmul(((matmul(keys, input.T) - input.sum) / temperature).T, values)
 
 // Returns a value (into the output matrix) given the input key from the hopfield dictionary.
-void hopfield_dictionary_get(Eigen::MatrixXd & input, Eigen::MatrixXd & keys, Eigen::MatrixXd & values, Eigen::MatrixXd & out, double temperature){
+void hopfield_dictionary_get(Eigen::MatrixXf & input, Eigen::MatrixXf & keys, Eigen::MatrixXf & values, Eigen::MatrixXf & out, float temperature){
     out = ((keys * input.transpose() - input.rowwise().sum().transpose().replicate(keys.rows(),1)).array() / temperature).exp().matrix().transpose() * values;
 }
